requirements.txt
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
Flask==2.2.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
pytest==7.2.2
python-dotenv==1.0.0
requests==2.28.2
tomli==2.0.1
urllib3==1.26.14
Werkzeug==2.2.2
